requirements.txt

astroid==3.0.1
backports.entry-points-selectable==1.1.1
branca==0.6.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
clikit==0.6.2
colorama==0.4.6
coverage==7.3.2
crashtest==0.3.1
cryptography==35.0.0
debugpy==1.2.1
dill==0.3.7
distlib==0.3.3

django-phonenumber-field==7.2.0
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.3
filelock==3.3.2
folium==0.14.0
future==0.18.2
geographiclib==2.0
geopy==2.4.0
html5lib==1.1
idna==3.3
importlib-metadata==3.10.1
iniconfig==2.0.0
isort==5.12.0
jedi==0.15.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==21.8.0
lockfile==0.12.2
mando==0.7.1
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.2

numpy==1.24.4
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
phonenumberslite==8.13.24
Pillow==10.1.0
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
poetry==1.1.11
poetry-core==1.0.7
ptyprocess==0.7.0
pycparser==2.21
pyflakes==2.1.1
pylev==1.4.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.0.5
pytest==7.4.3
pytest-django==4.7.0
python-jsonrpc-server==0.3.2
python-language-server==0.31.10

radon==6.0.1
requests==2.26.0
requests-toolbelt==0.9.1
rope==0.18.0
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0

toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
typing-extensions==4.8.0
ujson==1.35
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
yapf==0.31.0
yarg==0.1.9
zipp==3.4.1
